vidutils: Scripts to automate some video editing steps through ffmpeg
This package will install the following scripts that rely on ffmpeg:

vid-split:
vid-volume:
vid-merge:

The edits these scripts perform are fast when done through ffmpeg
compared to the freely available graphical video editors, however the
ffmpeg filter syntax is difficult to learn and assembling the filters
manually is time-consuming. The vidutils scripts automate assembling and
running the filters. Full details are in the README.md file.